build_image/check_diff: update origin/master
Before, origin/master was not updated. To still check the
diff against the correct revision on Jenkins the Git repo
had been removed and re-cloned again. This is for one
particular reason very bad; it prevents `docker build`
from using cached intermediate images.

Now, we don't need to touch the repo anymore and the CI
job on Jenkins should have a noticeable boost.

Change-Id: Ibfea1509964c578c4d69bbbd2d59b75b2eeeca10